+++ src/components/header/header.test.tsx
@@ -20,4 +20,84 @@ describe('EuiHeader', () => {
 
     expect(component).toMatchSnapshot();
   });
+
+  test('renders in fixed position', () => {
+    const component = render(
+      <EuiHeader position="fixed">
+        <span>Hello!</span>
+      </EuiHeader>
+    );
+
+    expect(component).toMatchSnapshot();
+  });
+
+  describe('sections', () => {
+    test('render simple items and borders', () => {
+      const component = render(
+        <EuiHeader
+          sections={[
+            {
+              items: ['Item 1', 'Item 2'],
+              borders: 'right',
+            },
+            {
+              items: ['Item A', 'Item B'],
+            },
+          ]}
+        />
+      );
+
+      expect(component).toMatchSnapshot();
+    });
+
+    test('render breadcrumbs and props', () => {
+      const component = render(
+        <EuiHeader
+          sections={[
+            {
+              breadcrumbs: [{ text: 'Breadcrumb' }],
+              breadcrumbProps: { responsive: false },
+            },
+          ]}
+        />
+      );
+
+      expect(component).toMatchSnapshot();
+    });
+  });
+
+  describe('throws a warning', () => {
+    const oldConsoleError = console.warn;
+    let consoleStub: jest.Mock;
+
+    beforeEach(() => {
+      // We don't use jest.spyOn() here, because EUI's tests apply a global
+      // console.error() override that throws an exception. For these
+      // tests, we just want to know if console.error() was called.
+      console.warn = consoleStub = jest.fn();
+    });
+
+    afterEach(() => {
+      console.warn = oldConsoleError;
+    });
+
+    test('if both children and sections were passed', () => {
+      const component = render(
+        <EuiHeader
+          sections={[
+            {
+              items: ['Item 1', 'Item 2'],
+            },
+          ]}>
+          Child
+        </EuiHeader>
+      );
+
+      expect(consoleStub).toBeCalled();
+      expect(consoleStub.mock.calls[0][0]).toMatch(
+        'cannot accept both `children` and `sections`'
+      );
+      expect(component).toMatchSnapshot();
+    });
+  });
 });